pyproject.toml
[build-system]
requires = ["uv_build >= 0.10.8, <0.11.0"]
build-backend = "uv_build"

[tool.setuptools.package-dir]
"" = "src"

[project]
name = "rag"
version = "0.1.0"
authors = [
  {name = "Roman Drovskikh", email = "dorovskih2005@gmail.com"},
]
requires-python = "==3.11.*"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
dependencies = [
    "black>=26.3.1",
    "docling>=2.80.0",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "mypy>=1.19.1",
    "openai>=2.26.0",
    "pydantic-settings>=2.13.1",
    "pypdf>=6.8.0",
    "ruff>=0.15.5",
    "torch",
    "torchvision",
    "torchaudio",
    "qdrant-client>=1.17.1",
    "langchain-text-splitters>=1.1.1",
    "fastapi[standard]>=0.135.1",
]

[tool.uv]
index-strategy = "unsafe-best-match"

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu126" }
torchvision = { index = "pytorch-cu126" }
torchaudio = { index = "pytorch-cu126" }

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "C90"]
ignore = []

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = "src"
explicit_package_bases = true
namespace_packages = true
ignore_missing_imports = true

[tool.fastapi]
entrypoint = "rag.backend.main:app"